WC-2728: products.cloud.model: "directory"-like → graph
The data structure WideSky presents is a directed graph.  A
"directory"-like structure would be a tree, which is a graph that does
not contain cycles.

WideSky cannot be described as a tree as the references can contain
cycles.  The linked Wikipedia article provides the mathematical
background on what this represents.